fix: move gravity precompile registration to rpc impl
