+++ crates/rpc/rpc/Cargo.toml
@@ -40,6 +40,7 @@ reth-consensus.workspace = true
 reth-consensus-common.workspace = true
 reth-node-api.workspace = true
 reth-trie-common.workspace = true
+gravity-precompiles.workspace = true
 
 # ethereum
 alloy-evm = { workspace = true, features = ["overrides"] }
