+++ crates/rpc/rpc/src/eth/helpers/call.rs
@@ -1,13 +1,47 @@
 //! Contains RPC handler implementations specific to endpoints that call/execute within evm.
 
 use crate::EthApi;
-use reth_evm::{SpecFor, TxEnvFor};
+use alloy_consensus::BlockHeader;
+use alloy_primitives::{B256, U256};
+use gravity_precompiles::randomness_by_height::{
+    create_randomness_by_height_precompile, RandomnessByHeightProvider,
+    RANDOMNESS_BY_HEIGHT_PRECOMPILE_ADDR,
+};
+use reth_chainspec::{ChainSpecProvider, EthChainSpec, GravityHardfork};
+use reth_errors::ProviderError;
+use reth_evm::{precompiles::PrecompilesMap, Evm, SpecFor, TxEnvFor};
 use reth_rpc_convert::RpcConvert;
 use reth_rpc_eth_api::{
     helpers::{estimate::EstimateCall, Call, EthCall},
     FromEvmError, RpcNodeCore,
 };
 use reth_rpc_eth_types::EthApiError;
+use reth_storage_api::HeaderProvider;
+use std::sync::Arc;
+
+#[derive(Clone, Debug)]
+struct HeaderRandomnessProvider<Provider> {
+    provider: Provider,
+}
+
+impl<Provider> HeaderRandomnessProvider<Provider> {
+    const fn new(provider: Provider) -> Self {
+        Self { provider }
+    }
+}
+
+impl<Provider> RandomnessByHeightProvider for HeaderRandomnessProvider<Provider>
+where
+    Provider: HeaderProvider,
+{
+    type Error = ProviderError;
+
+    fn randomness_by_height(&self, height: u64) -> Result<Option<B256>, Self::Error> {
+        self.provider
+            .header_by_number(height)
+            .map(|header| header.and_then(|header| header.mix_hash()))
+    }
+}
 
 impl<N, Rpc> EthCall for EthApi<N, Rpc>
 where
@@ -41,6 +75,27 @@ where
     #[inline]
     fn max_simulate_blocks(&self) -> u64 {
         self.inner.max_simulate_blocks()
+    }
+
+    fn register_custom_precompiles<EV>(&self, evm: &mut EV, block_number: U256)
+    where
+        EV: Evm<Precompiles = PrecompilesMap>,
+    {
+        let Ok(block_number) = u64::try_from(block_number) else { return };
+        if !self
+            .provider()
+            .chain_spec()
+            .gravity_hardforks()
+            .is_fork_active_at_block(GravityHardfork::Alpha, block_number)
+        {
+            return
+        }
+
+        let precompile = create_randomness_by_height_precompile(Arc::new(
+            HeaderRandomnessProvider::new(self.provider().clone()),
+        ));
+        evm.precompiles_mut()
+            .apply_precompile(&RANDOMNESS_BY_HEIGHT_PRECOMPILE_ADDR, move |_| Some(precompile));
     }
 }
